my_app/serializers/subtask.py
from rest_framework import serializers
from my_app.models import SubTask




class SubTaskSerializer(serializers.ModelSerializer):
    class Meta:
        model = SubTask
        fields = [
            'title',
            'description',
            'deadline',
            'created_at',
        ]


class SubTaskCreateSerializer(serializers.ModelSerializer):
    created_at = serializers.DateTimeField(read_only=True)

    class Meta:
        model = SubTask
        fields = "__all__"