requirements.txt
asgiref==3.11.1
Django==6.0.2
django-environ==0.13.0
sqlparse==0.5.5
tzdata==2025.3

djangorestframework